+++ Clicked/Locales/enUS.lua
@@ -308,3 +308,4 @@ L["Always"] = true
 L["Automatically import from macros"] = true
 L["Are you sure you want to delete these %d bindings and/or groups?"] = true
 L["copy"] = true
+L["Action bar page"] = true